moving the functions to rgw_orphan_list in reusable folder
